03-Bert_Phi3_Rag_Chat/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.42.4
numpy==1.26.4
torch==2.3.1+cu121